refactor: extract resuable V3 views walker
