+++ core/views.py
@@ -23,7 +23,7 @@ from django.http import (
 )
 from django.shortcuts import redirect
 from django.template.loader import render_to_string
-from django.urls import URLPattern, URLResolver, get_resolver, reverse
+from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.views import View
 from django.views.decorators.cache import never_cache
@@ -42,7 +42,7 @@ from libraries.utils import (
 )
 from versions.models import Version, docs_path_to_boost_name
 
-from .mixins import V3Mixin
+from .mixins import V3Mixin, iter_v3_views
 from .asciidoc import convert_adoc_to_html
 from .boostrenderer import (
     convert_img_paths,
@@ -1908,34 +1908,16 @@ class V3ComponentDemoView(V3Mixin, TemplateView):
         context["demo_library_items"] = demo_library_items
 
         # V3 paths registry
-        v3_paths = []
-
-        def walk(patterns):
-            for entry in patterns:
-                if isinstance(entry, URLResolver):
-                    walk(entry.url_patterns)
-                elif isinstance(entry, URLPattern):
-                    callback = entry.callback
-                    while callback is not None:
-                        view_class = getattr(callback, "view_class", None)
-                        if view_class is not None:
-                            break
-                        callback = getattr(callback, "__wrapped__", None)
-                    if view_class and issubclass(view_class, V3Mixin):
-                        has_legacy = getattr(view_class, "template_name", None)
-                        v3_paths.append(
-                            {
-                                "class_name": view_class.__name__,
-                                "url": f"/{entry.pattern}",
-                                "url_name": entry.name or "",
-                                "v3_template": getattr(
-                                    view_class, "v3_template_name", ""
-                                ),
-                                "has_legacy_template": bool(has_legacy),
-                            }
-                        )
-
-        walk(get_resolver().url_patterns)
+        v3_paths = [
+            {
+                "class_name": view_class.__name__,
+                "url": f"/{entry.pattern}",
+                "url_name": entry.name or "",
+                "v3_template": getattr(view_class, "v3_template_name", ""),
+                "has_legacy_template": bool(getattr(view_class, "template_name", None)),
+            }
+            for entry, view_class in iter_v3_views()
+        ]
         v3_paths.sort(key=lambda p: p["class_name"])
         context["v3_paths"] = v3_paths
